+++ components/ui/Input/index.tsx
@@ -1,2 +1,2 @@
-export { default as TextInput, InputStyle } from './TextInput';
+export { default as Input, InputStyle } from './Input';
 export { default as TextArea } from './TextArea';
